Migrate `RequestLocationPermission` to Mobius
